regularly check the form's inputFields
used window.setInterval(x, interval) to regularly check the form's fields and see if they are all filled in in order to enable the submit button. window.setInterval() will run every 500ms to check on the contents of the form.

Updated the checkInputField() function to checkInputField(inputField) to allow an input field to be passed through as a parameter. Doing so allows the logic in the function to be much more reusable as it can check the current value of the input field passed through. Along with that the function will return a boolean value depending on whether or not the field is empty or not