added email validation
if the email address is in a valid format and is a registered one an alert will pop up showing what the password for the account was. If the email is in a valid format but not registered an alert banner will just show up indicating an email has been sent.